Hop N Hop
Big Bunny lives in a colony. He is the only bunny in his colony who is not able to hop. On his 5th birthday, his father bunny gifted him a pogo stick as he could not jump like the other bunnies. He is so excited to play with the pogo stick. The pogo stick hops one unit per jump. He wanders around his house jumping with pogo sticks. He wants to show the pogo stick to his friends and decides to go using pogo sticks. Write a program to find the number of hops needed to reach his friends' house. Assume that Big Bunny's house is in the location (3,4).
Input format:-
Input consists of two integers x,y.
The x and y corresponds to x and y coordinates of his friend's house.

Output format: -
The output is an integer. It corresponds to the hops needs to reach his friend's house.
Sample Input:-
5
10
Sample output:-
6